+++ packages/liveblocks-core/src/index.ts
@@ -233,6 +233,7 @@ export type {
 } from "./protocol/Comments";
 export type { QueryMetadata } from "./protocol/Comments";
 export type {
+  SearchCommentsResult,
   ThreadData,
   ThreadDataPlain,
   ThreadDataWithDeleteInfo,
@@ -298,7 +299,6 @@ export type {
   SerializedRootObject,
 } from "./protocol/SerializedCrdt";
 export { CrdtType } from "./protocol/SerializedCrdt";
-export { isChildCrdt, isRootCrdt } from "./protocol/SerializedCrdt";
 export type {
   BroadcastedEventServerMsg,
   CommentsEventServerMsg,