+++ packages/liveblocks-core/src/protocol/Comments.ts
@@ -154,6 +154,12 @@ export type CommentUserReaction = {
 
 export type CommentUserReactionPlain = DateToString<CommentUserReaction>;
 
+export type SearchCommentsResult = {
+  threadId: string;
+  commentId: string;
+  content: string;
+};
+
 /**
  * Represents a thread of comments.
  */