+++ packages/liveblocks-node/src/client.ts
@@ -154,7 +154,8 @@ export type CreateThreadOptions<M extends BaseMetadata> = {
 export type RoomPermission =
   | []
   | ["room:write"]
-  | ["room:read", "room:presence:write"];
+  | ["room:read", "room:presence:write"]
+  | ["room:read", "room:presence:write", "comments:write"];
 export type RoomAccesses = Record<
   string,
   | ["room:write"]
@@ -205,7 +206,10 @@ type OpenAiModel =
   | "gpt-5"
   | "gpt-5-mini"
   | "gpt-5-nano"
-  | "gpt-5-chat-latest";
+  | "gpt-5-chat-latest"
+  | "gpt-5.1"
+  | "gpt-5.1-mini"
+  | "gpt-5.1-chat-latest";
 
 type OpenAiProviderOptions = {
   openai: {